Add support to dynamically load controllters
This allows controllers to be loaded when they're called. One issue is that if a
user names a controller `some_controller/rb` but th file contains
AnotherController then an endless loop of `Object.const_get` will occur which is
why there's a pretty hacky way added to determine if `#cont_get` has been called
for the specific object.

Here's how rails does it -
https://github.com/rails/rails/blob/0ed6ebcf9095c65330d3950cfb6b75ba7ea78853/activesupport/lib/active_support/dependencies.rb